Update tests & TestDatabase class
