+++ integrations/acquisition/covid_hosp/state_timeseries/test_scenarios.py
@@ -57,7 +57,7 @@ class AcquisitionTests(unittest.TestCase):
     with self.subTest(name='first acquisition'), \
          patch.object(Network, 'fetch_metadata', return_value=self.test_utils.load_sample_metadata()), \
          patch.object(Network, 'fetch_dataset', return_value=self.test_utils.load_sample_dataset()):
-      acquired = Utils.update_dataset(Database)
+      acquired = Database().update_dataset()
       self.assertTrue(acquired)
 
     # make sure the data now exists
@@ -83,7 +83,7 @@ class AcquisitionTests(unittest.TestCase):
     with self.subTest(name='second acquisition'), \
          patch.object(Network, 'fetch_metadata', return_value=self.test_utils.load_sample_metadata()), \
          patch.object(Network, 'fetch_dataset', return_value=self.test_utils.load_sample_dataset()):
-      acquired = Utils.update_dataset(Database)
+      acquired = Database().update_dataset()
       self.assertFalse(acquired)
 
     # make sure the data still exists
@@ -97,7 +97,7 @@ class AcquisitionTests(unittest.TestCase):
          patch.object(Network, 'fetch_metadata', return_value=self.test_utils.load_sample_metadata("metadata2.csv")), \
          patch.object(Network, 'fetch_dataset', return_value=self.test_utils.load_sample_dataset("dataset2.csv")):
       # acquire new data with 3/16 issue date
-      acquired = Utils.update_dataset(Database)
+      acquired = Database().update_dataset()
       self.assertTrue(acquired)
 
     with self.subTest(name='as_of checks'):
