+++ tests/acquisition/covid_hosp/common/test_utils.py
@@ -1,14 +1,12 @@
 """Unit tests for utils.py."""
 
 # standard library
-from datetime import date
 import unittest
-from unittest.mock import MagicMock, PropertyMock, patch
+from unittest.mock import MagicMock, patch
 
 # first party
-from delphi.epidata.acquisition.covid_hosp.common.database import Database
 from delphi.epidata.acquisition.covid_hosp.common.network import Network
-from delphi.epidata.acquisition.covid_hosp.common.test_utils import UnitTestUtils
+from delphi.epidata.acquisition.covid_hosp.common.test_utils import TestDatabase, UnitTestUtils
 from delphi.epidata.acquisition.covid_hosp.common.utils import Utils, CovidHospException
 from delphi.epidata.common.covid_hosp.covid_hosp_schema_io import CovidHospSomething
 
@@ -18,9 +16,6 @@ import mysql.connector
 
 # py3tester coverage target
 __test_target__ = 'delphi.epidata.acquisition.covid_hosp.common.utils'
-
-class TestDatabase(Database):
-  DATASET_NAME = 'mock_dataset'
 
 class UtilsTests(unittest.TestCase):
 
@@ -59,26 +54,6 @@ class UtilsTests(unittest.TestCase):
     with self.subTest(name='exception'):
       with self.assertRaises(CovidHospException):
         Utils.parse_bool('maybe')
-
-  def test_issues_to_fetch(self):
-    test_metadata = pd.DataFrame({
-      "date": [pd.Timestamp("2021-03-13 00:00:00"),
-               pd.Timestamp("2021-03-14 00:00:00"),
-               pd.Timestamp("2021-03-15 00:00:01"),
-               pd.Timestamp("2021-03-15 00:00:00"),
-               pd.Timestamp("2021-03-16 00:00:00")
-               ],
-      "Archive Link": ["a", "b", "d", "c", "e"]
-    }).set_index("date")
-
-    issues = Utils.issues_to_fetch(test_metadata, pd.Timestamp("2021-3-13"), pd.Timestamp("2021-3-16"))
-    self.assertEqual(issues,
-                     {date(2021, 3, 14): [("b", pd.Timestamp("2021-03-14 00:00:00"))],
-                      date(2021, 3, 15): [("c", pd.Timestamp("2021-03-15 00:00:00")),
-                                          ("d", pd.Timestamp("2021-03-15 00:00:01"))]
-                      }
-                     )
-
   def test_run_skip_old_dataset(self):
     """Don't re-acquire an old dataset."""
 
@@ -96,7 +71,7 @@ class UtilsTests(unittest.TestCase):
          patch.object(TestDatabase, 'get_max_issue', return_value=pd.Timestamp("2200/1/1")), \
          patch.object(TestDatabase, 'insert_metadata', return_value=None) as insert_metadata, \
          patch.object(TestDatabase, 'insert_dataset', return_value=None) as insert_dataset:
-      result = Utils.update_dataset(database=TestDatabase)
+      result = TestDatabase().update_dataset()
 
     self.assertFalse(result)
     fetch_dataset.assert_not_called()
@@ -111,8 +86,7 @@ class UtilsTests(unittest.TestCase):
     fake_issues = {pd.Timestamp("2021/3/15"): [("url1", pd.Timestamp("2021-03-15 00:00:00")),
                                               ("url2", pd.Timestamp("2021-03-15 00:00:00"))]}
 
-    with patch.object(Utils, 'issues_to_fetch', return_value=fake_issues), \
-         patch.object(Network, 'fetch_metadata', return_value=self.test_utils.load_sample_metadata()), \
+    with patch.object(Network, 'fetch_metadata', return_value=self.test_utils.load_sample_metadata()), \
          patch.object(Network, 'fetch_dataset', return_value=fake_dataset), \
          patch.object(CovidHospSomething, 'get_ds_table_name', return_value=None), \
          patch.object(CovidHospSomething, 'get_ds_dataset_id', return_value=None), \
@@ -123,9 +97,10 @@ class UtilsTests(unittest.TestCase):
          patch.object(CovidHospSomething, 'get_ds_aggregate_key_cols', return_value=None), \
          patch.object(mysql.connector, 'connect', return_value=mock_connection), \
          patch.object(TestDatabase, 'get_max_issue', return_value=pd.Timestamp("1900/1/1")), \
+         patch.object(TestDatabase, 'issues_to_fetch', return_value=fake_issues), \
          patch.object(TestDatabase, 'insert_metadata', return_value=None) as insert_metadata, \
          patch.object(TestDatabase, 'insert_dataset', return_value=None) as insert_dataset:
-        result = Utils.update_dataset(database=TestDatabase)
+        result = TestDatabase().update_dataset()
 
     self.assertTrue(result)
